Now Uploaded on Live server

+++ protected/modules/admin/views/loan/_view_loans.php
@@ -12,6 +12,9 @@
 	<td>
 		<?php
 			switch ($data->status) {
+				case 'I':
+					echo "<label class='label label-info'>Approved - Invitation</label>";
+					break;
 				case 'A':
 					echo "<label class='label label-success'>Approved</label>";
 					break;
@@ -34,23 +37,25 @@
             <li>
             	<?php echo CHtml::link('<span class="fa fa-search"></span> View Profile', 'javascript:void(0);', array('class'=>'view-account', 'data-toggle'=>'modal', 'data-target'=>'.bs-example-modal-lg', 'data-url'=>Yii::app()->createUrl('admin/account/view', array('id'=>$data->account_id)))); ?>
             </li>
-            <li class="divider"></li>
-            <li>
-	            <?php
-					switch ($data->status) {
-						case 'P':
-							echo CHtml::link('<span class="fa fa-check"></span> Approve', array('loan/approve', 'id'=>$data->id), array('confirm'=>'Are you sure you want to approve this application?', 'title'=>'Approve Investment'));
-							echo CHtml::link('<span class="fa fa-times"></span> Reject', array('loan/reject', 'id'=>$data->id), array('confirm'=>'Are you sure you want to reject this application?', 'title'=>'Reject Investment'));
-							break;
-						case 'A':
-							echo CHtml::link('<span class="fa fa-times"></span> Reject', array('loan/reject', 'id'=>$data->id), array('confirm'=>'Are you sure you want to reject this application?', 'title'=>'Reject Investment'));
-							break;
-						case 'R':
-							echo CHtml::link('<span class="fa fa-check"></span> Approve', array('loan/approve', 'id'=>$data->id), array('confirm'=>'Are you sure you want to approve this application?', 'title'=>'Approve Investment'));
-							break;
-					}
-				?>
-            </li>
+            <?php if ($data->status != "I"): ?>
+            	<li class="divider"></li>
+	            <li>
+		            <?php
+						switch ($data->status) {
+							case 'P':
+								echo CHtml::link('<span class="fa fa-check"></span> Approve', array('loan/approve', 'id'=>$data->id), array('confirm'=>'Are you sure you want to approve this application?', 'title'=>'Approve Investment'));
+								echo CHtml::link('<span class="fa fa-times"></span> Reject', array('loan/reject', 'id'=>$data->id), array('confirm'=>'Are you sure you want to reject this application?', 'title'=>'Reject Investment'));
+								break;
+							case 'A':
+								echo CHtml::link('<span class="fa fa-times"></span> Reject', array('loan/reject', 'id'=>$data->id), array('confirm'=>'Are you sure you want to reject this application?', 'title'=>'Reject Investment'));
+								break;
+							case 'R':
+								echo CHtml::link('<span class="fa fa-check"></span> Approve', array('loan/approve', 'id'=>$data->id), array('confirm'=>'Are you sure you want to approve this application?', 'title'=>'Approve Investment'));
+								break;
+						}
+					?>
+	            </li>
+            <?php endif ?>
           </ul>
     	</div>
 	</td>
